Per system inboxes
This fixes an unintuitive situation where an inbox appears in the
system list before the outbox. In that case, events are sent, but
never received because all the mailboxes were cleared at the end
of every tick.

With this change, each system gets its own inbox, which is cleared
after the system executes. It will mean more memory usage, but the
difference should be small. And the behavior will be much more
intuitive.